fix add document

+++ src/front/myb.front/libs/doc-management-module/src/lib/components/Folders/folder-cards/folder-cards.component.ts
@@ -63,6 +63,9 @@ documents: DocumentModel[]=[];
 
   //delete doc
 deleteFolder(id: number){
+  if (id==-1) {
+    return;
+  }
   this.folderService.delete(id).subscribe(() => {
     this.loadFolders();
     // this.documents = this.documents.filter(doc => doc.id !== docId);
